move filter to inside role
