+++ src/commcare_cloud/ansible/ansible.cfg
@@ -2,7 +2,6 @@
 ansible_managed = Ansible managed, do not edit directly
 log_path=/var/log/ansible.log
 callback_plugins = ./plugins/callback
-filter_plugins = ./plugins/filter
 callback_whitelist = datadog_callback
 
 gathering = smart
